+++ src/ui/map.test.ts
@@ -76,34 +76,6 @@ describe('Map', () => {
         //t.error();
     });
 
-    test('initial bounds in constructor options', () => {
-        const container = window.document.createElement('div');
-        Object.defineProperty(container, 'offsetWidth', {value: 512});
-        Object.defineProperty(container, 'offsetHeight', {value: 512});
-
-        const bounds = [[-133, 16], [-68, 50]];
-        const map = createMap({container, bounds});
-
-        expect(fixedLngLat(map.getCenter(), 4)).toEqual({lng: -100.5, lat: 34.7171});
-        expect(fixedNum(map.getZoom(), 3)).toBe(2.113);
-    });
-
-    test('initial bounds options in constructor options', () => {
-        const bounds = [[-133, 16], [-68, 50]];
-
-        const map = (fitBoundsOptions) => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'offsetWidth', {value: 512});
-            Object.defineProperty(container, 'offsetHeight', {value: 512});
-            return createMap({container, bounds, fitBoundsOptions});
-        };
-
-        const unpadded = map(undefined);
-        const padded = map({padding: 100});
-
-        expect(unpadded.getZoom() > padded.getZoom()).toBeTruthy();
-    });
-
     describe('disables handlers', () => {
         test('disables all handlers', () => {
             const map = createMap({interactive: false});
@@ -153,55 +125,6 @@ describe('Map', () => {
             map.setStyle(createStyle());
         }, 1);
 
-    });
-
-    describe('#mapOptions', () => {
-        test('maxTileCacheZoomLevels: Default value is set', () => {
-            const map = createMap();
-            expect(map._maxTileCacheZoomLevels).toBe(config.MAX_TILE_CACHE_ZOOM_LEVELS);
-        });
-
-        test('maxTileCacheZoomLevels: Value can be set via map options', () => {
-            const map = createMap({maxTileCacheZoomLevels: 1});
-            expect(map._maxTileCacheZoomLevels).toBe(1);
-        });
-
-        test('Style validation is enabled by default', () => {
-            let validationOption = false;
-            jest.spyOn(Style.prototype, 'loadJSON').mockImplementationOnce((styleJson, options) => {
-                validationOption = options.validate;
-            });
-            createMap();
-            expect(validationOption).toBeTruthy();
-        });
-
-        test('Style validation disabled using mapOptions', () => {
-            let validationOption = true;
-            jest.spyOn(Style.prototype, 'loadJSON').mockImplementationOnce((styleJson, options) => {
-                validationOption = options.validate;
-            });
-            createMap({validateStyle: false});
-
-            expect(validationOption).toBeFalsy();
-        });
-
-        test('fadeDuration is set after first idle event', async () => {
-            let idleTriggered = false;
-            const fadeDuration = 100;
-            const spy = jest.spyOn(Style.prototype, 'update').mockImplementation((parameters: EvaluationParameters) => {
-                if (!idleTriggered) {
-                    expect(parameters.fadeDuration).toBe(0);
-                } else {
-                    expect(parameters.fadeDuration).toBe(fadeDuration);
-                }
-            });
-            const style = createStyle();
-            const map = createMap({style, fadeDuration});
-            await map.once('idle');
-            idleTriggered = true;
-            map.zoomTo(0.5, {duration: 100});
-            spy.mockRestore();
-        });
     });
 
     describe('#setTransformRequest', () => {
@@ -286,142 +209,6 @@ describe('Map', () => {
                 done();
             });
         });
-    });
-
-    describe('#getStyle', () => {
-        test('returns undefined if the style has not loaded yet', done => {
-            const style = createStyle();
-            const map = createMap({style});
-            expect(map.getStyle()).toBeUndefined();
-            done();
-        });
-
-        test('returns the style', done => {
-            const style = createStyle();
-            const map = createMap({style});
-
-            map.on('load', () => {
-                expect(map.getStyle()).toEqual(style);
-                done();
-            });
-        });
-
-        test('returns the style with added sources', done => {
-            const style = createStyle();
-            const map = createMap({style});
-
-            map.on('load', () => {
-                map.addSource('geojson', createStyleSource());
-                expect(map.getStyle()).toEqual(extend(createStyle(), {
-                    sources: {geojson: createStyleSource()}
-                }));
-                done();
-            });
-        });
-
-        test('fires an error on checking if non-existant source is loaded', done => {
-            const style = createStyle();
-            const map = createMap({style});
-
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/There is no source with ID/);
-                    done();
-                });
-                map.isSourceLoaded('geojson');
-            });
-        });
-
-        test('returns the style with added layers', done => {
-            const style = createStyle();
-            const map = createMap({style});
-            const layer = {
-                id: 'background',
-                type: 'background'
-            } as LayerSpecification;
-
-            map.on('load', () => {
-                map.addLayer(layer);
-                expect(map.getStyle()).toEqual(extend(createStyle(), {
-                    layers: [layer]
-                }));
-                done();
-            });
-        });
-
-        test('a layer can be added even if a map is created without a style', () => {
-            const map = createMap({deleteStyle: true});
-            const layer = {
-                id: 'background',
-                type: 'background'
-            } as LayerSpecification;
-            map.addLayer(layer);
-        });
-
-        test('a source can be added even if a map is created without a style', () => {
-            const map = createMap({deleteStyle: true});
-            const source = createStyleSource();
-            map.addSource('fill', source);
-        });
-
-        test('a layer can be added with an embedded source specification', () => {
-            const map = createMap({deleteStyle: true});
-            const source: GeoJSONSourceSpecification = {
-                type: 'geojson',
-                data: {type: 'Point', coordinates: [0, 0]}
-            };
-            map.addLayer({
-                id: 'foo',
-                type: 'symbol',
-                source
-            });
-        });
-
-        test('returns the style with added source and layer', done => {
-            const style = createStyle();
-            const map = createMap({style});
-            const source = createStyleSource();
-            const layer = {
-                id: 'fill',
-                type: 'fill',
-                source: 'fill'
-            } as LayerSpecification;
-
-            map.on('load', () => {
-                map.addSource('fill', source);
-                map.addLayer(layer);
-                expect(map.getStyle()).toEqual(extend(createStyle(), {
-                    sources: {fill: source},
-                    layers: [layer]
-                }));
-                done();
-            });
-        });
-
-        test('creates a new Style if diff fails', () => {
-            const style = createStyle();
-            const map = createMap({style});
-            jest.spyOn(map.style, 'setState').mockImplementation(() => {
-                throw new Error('Dummy error');
-            });
-            jest.spyOn(console, 'warn').mockImplementation(() => {});
-
-            const previousStyle = map.style;
-            map.setStyle(style);
-            expect(map.style && map.style !== previousStyle).toBeTruthy();
-        });
-
-        test('creates a new Style if diff option is false', () => {
-            const style = createStyle();
-            const map = createMap({style});
-            const spy = jest.spyOn(map.style, 'setState');
-
-            const previousStyle = map.style;
-            map.setStyle(style, {diff: false});
-            expect(map.style && map.style !== previousStyle).toBeTruthy();
-            expect(spy).not.toHaveBeenCalled();
-        });
-
     });
 
     test('#moveLayer', async () => {
@@ -510,298 +297,6 @@ describe('Map', () => {
                 expect(map.getLayersOrder()).toEqual(['custom', 'raster']);
                 done();
             });
-        });
-    });
-
-    describe('#resize', () => {
-        test('sets width and height from container clients', () => {
-            const map = createMap(),
-                container = map.getContainer();
-
-            Object.defineProperty(container, 'clientWidth', {value: 250});
-            Object.defineProperty(container, 'clientHeight', {value: 250});
-            map.resize();
-
-            expect(map.transform.width).toBe(250);
-            expect(map.transform.height).toBe(250);
-
-        });
-
-        test('fires movestart, move, resize, and moveend events', () => {
-            const map = createMap(),
-                events = [];
-
-            (['movestart', 'move', 'resize', 'moveend'] as any).forEach((event) => {
-                map.on(event, (e) => {
-                    events.push(e.type);
-                });
-            });
-
-            map.resize();
-            expect(events).toEqual(['movestart', 'move', 'resize', 'moveend']);
-
-        });
-
-        test('listen to window resize event', () => {
-            const spy = jest.fn();
-            global.ResizeObserver = jest.fn().mockImplementation(() => ({
-                observe: spy
-            }));
-
-            createMap();
-
-            expect(spy).toHaveBeenCalled();
-        });
-
-        test('do not resize if trackResize is false', () => {
-            let observerCallback: Function = null;
-            global.ResizeObserver = jest.fn().mockImplementation((c) => ({
-                observe: () => { observerCallback = c; }
-            }));
-
-            const map = createMap({trackResize: false});
-
-            const spyA = jest.spyOn(map, 'stop');
-            const spyB = jest.spyOn(map, '_update');
-            const spyC = jest.spyOn(map, 'resize');
-
-            observerCallback();
-
-            expect(spyA).not.toHaveBeenCalled();
-            expect(spyB).not.toHaveBeenCalled();
-            expect(spyC).not.toHaveBeenCalled();
-        });
-
-        test('do resize if trackResize is true (default)', async () => {
-            let observerCallback: Function = null;
-            global.ResizeObserver = jest.fn().mockImplementation((c) => ({
-                observe: () => { observerCallback = c; }
-            }));
-
-            const map = createMap();
-
-            const updateSpy = jest.spyOn(map, '_update');
-            const resizeSpy = jest.spyOn(map, 'resize');
-
-            // The initial "observe" event fired by ResizeObserver should be captured/muted
-            // in the map constructor
-
-            observerCallback();
-            expect(updateSpy).not.toHaveBeenCalled();
-            expect(resizeSpy).not.toHaveBeenCalled();
-
-            // The next "observe" event should fire a resize / _update
-
-            observerCallback();
-            expect(updateSpy).toHaveBeenCalled();
-            expect(resizeSpy).toHaveBeenCalledTimes(1);
-
-            // Additional "observe" events should be throttled
-            observerCallback();
-            observerCallback();
-            observerCallback();
-            observerCallback();
-            expect(resizeSpy).toHaveBeenCalledTimes(1);
-            await sleep(100);
-            expect(resizeSpy).toHaveBeenCalledTimes(2);
-        });
-
-        test('width and height correctly rounded', () => {
-            const map = createMap();
-            const container = map.getContainer();
-
-            Object.defineProperty(container, 'clientWidth', {value: 250.6});
-            Object.defineProperty(container, 'clientHeight', {value: 250.6});
-            map.resize();
-
-            expect(map.getCanvas().width).toBe(250);
-            expect(map.getCanvas().height).toBe(250);
-            expect(map.painter.width).toBe(250);
-            expect(map.painter.height).toBe(250);
-        });
-    });
-
-    describe('#getBounds', () => {
-
-        test('getBounds', () => {
-            const map = createMap({zoom: 0});
-            expect(parseFloat(map.getBounds().getCenter().lng.toFixed(10))).toBe(-0);
-            expect(parseFloat(map.getBounds().getCenter().lat.toFixed(10))).toBe(0);
-
-            expect(toFixed(map.getBounds().toArray())).toEqual(toFixed([
-                [-70.31249999999976, -57.326521225216965],
-                [70.31249999999977, 57.32652122521695]]));
-        });
-
-        test('rotated bounds', () => {
-            const map = createMap({zoom: 1, bearing: 45});
-            expect(
-                toFixed([[-49.718445552178764, -44.44541580601936], [49.7184455522, 44.445415806019355]])
-            ).toEqual(toFixed(map.getBounds().toArray()));
-
-            map.setBearing(135);
-            expect(
-                toFixed([[-49.718445552178764, -44.44541580601936], [49.7184455522, 44.445415806019355]])
-            ).toEqual(toFixed(map.getBounds().toArray()));
-
-        });
-
-        function toFixed(bounds) {
-            const n = 10;
-            return [
-                [normalizeFixed(bounds[0][0], n), normalizeFixed(bounds[0][1], n)],
-                [normalizeFixed(bounds[1][0], n), normalizeFixed(bounds[1][1], n)]
-            ];
-        }
-
-        function normalizeFixed(num, n) {
-            // workaround for "-0.0000000000" ≠ "0.0000000000"
-            return parseFloat(num.toFixed(n)).toFixed(n);
-        }
-    });
-
-    describe('#setMaxBounds', () => {
-        test('constrains map bounds', () => {
-            const map = createMap({zoom: 0});
-            map.setMaxBounds([[-130.4297, 50.0642], [-61.52344, 24.20688]]);
-            expect(
-                toFixed([[-130.4297000000, 7.0136641176], [-61.5234400000, 60.2398142283]])
-            ).toEqual(toFixed(map.getBounds().toArray()));
-        });
-
-        test('when no argument is passed, map bounds constraints are removed', () => {
-            const map = createMap({zoom: 0});
-            map.setMaxBounds([[-130.4297, 50.0642], [-61.52344, 24.20688]]);
-            expect(
-                toFixed([[-166.28906999999964, -27.6835270554], [-25.664070000000066, 73.8248206697]])
-            ).toEqual(toFixed(map.setMaxBounds(null).setZoom(0).getBounds().toArray()));
-        });
-
-        test('should not zoom out farther than bounds', () => {
-            const map = createMap();
-            map.setMaxBounds([[-130.4297, 50.0642], [-61.52344, 24.20688]]);
-            expect(map.setZoom(0).getZoom()).not.toBe(0);
-        });
-
-        function toFixed(bounds) {
-            const n = 9;
-            return [
-                [bounds[0][0].toFixed(n), bounds[0][1].toFixed(n)],
-                [bounds[1][0].toFixed(n), bounds[1][1].toFixed(n)]
-            ];
-        }
-
-    });
-
-    describe('#getMaxBounds', () => {
-        test('returns null when no bounds set', () => {
-            const map = createMap({zoom: 0});
-            expect(map.getMaxBounds()).toBeNull();
-        });
-
-        test('returns bounds', () => {
-            const map = createMap({zoom: 0});
-            const bounds = [[-130.4297, 50.0642], [-61.52344, 24.20688]] as LngLatBoundsLike;
-            map.setMaxBounds(bounds);
-            expect(map.getMaxBounds().toArray()).toEqual(bounds);
-        });
-
-    });
-
-    describe('#getRenderWorldCopies', () => {
-        test('initially false', () => {
-            const map = createMap({renderWorldCopies: false});
-            expect(map.getRenderWorldCopies()).toBe(false);
-        });
-
-        test('initially true', () => {
-            const map = createMap({renderWorldCopies: true});
-            expect(map.getRenderWorldCopies()).toBe(true);
-        });
-
-    });
-
-    describe('#setRenderWorldCopies', () => {
-        test('initially false', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.setRenderWorldCopies(true);
-            expect(map.getRenderWorldCopies()).toBe(true);
-        });
-
-        test('initially true', () => {
-            const map = createMap({renderWorldCopies: true});
-            map.setRenderWorldCopies(false);
-            expect(map.getRenderWorldCopies()).toBe(false);
-        });
-
-        test('undefined', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.setRenderWorldCopies(undefined);
-            expect(map.getRenderWorldCopies()).toBe(true);
-        });
-
-        test('null', () => {
-            const map = createMap({renderWorldCopies: true});
-            map.setRenderWorldCopies(null);
-            expect(map.getRenderWorldCopies()).toBe(false);
-        });
-
-    });
-
-    describe('renderWorldCopies', () => {
-        test('does not constrain horizontal panning when renderWorldCopies is set to true', () => {
-            const map = createMap({renderWorldCopies: true});
-            map.setCenter({lng: 180, lat: 0});
-            expect(map.getCenter().lng).toBe(180);
-        });
-
-        test('constrains horizontal panning when renderWorldCopies is set to false', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.setCenter({lng: 180, lat: 0});
-            expect(map.getCenter().lng).toBeCloseTo(110, 0);
-        });
-
-        test('does not wrap the map when renderWorldCopies is set to false', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.setCenter({lng: 200, lat: 0});
-            expect(map.getCenter().lng).toBeCloseTo(110, 0);
-        });
-
-        test('panTo is constrained to single globe when renderWorldCopies is set to false', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.panTo({lng: 180, lat: 0}, {duration: 0});
-            expect(map.getCenter().lng).toBeCloseTo(110, 0);
-            map.panTo({lng: -3000, lat: 0}, {duration: 0});
-            expect(map.getCenter().lng).toBeCloseTo(-110, 0);
-        });
-
-        test('flyTo is constrained to single globe when renderWorldCopies is set to false', () => {
-            const map = createMap({renderWorldCopies: false});
-            map.flyTo({center: [1000, 0], zoom: 3, animate: false});
-            expect(map.getCenter().lng).toBeCloseTo(171, 0);
-            map.flyTo({center: [-1000, 0], zoom: 5, animate: false});
-            expect(map.getCenter().lng).toBeCloseTo(-178, 0);
-        });
-
-        test('lng is constrained to a single globe when zooming with {renderWorldCopies: false}', () => {
-            const map = createMap({renderWorldCopies: false, center: [180, 0], zoom: 2});
-            expect(map.getCenter().lng).toBeCloseTo(162, 0);
-            map.zoomTo(1, {animate: false});
-            expect(map.getCenter().lng).toBeCloseTo(145, 0);
-        });
-
-        test('lng is constrained by maxBounds when {renderWorldCopies: false}', () => {
-            const map = createMap({
-                renderWorldCopies: false,
-                maxBounds: [
-                    [70, 30],
-                    [80, 40]
-                ],
-                zoom: 8,
-                center: [75, 35]
-            });
-            map.setCenter({lng: 180, lat: 0});
-            expect(map.getCenter().lng).toBeCloseTo(80, 0);
         });
     });
 
@@ -1519,425 +1014,6 @@ describe('Map', () => {
             });
         });
 
-    });
-
-    describe('#setFeatureState', () => {
-        test('sets state', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-                const fState = map.getFeatureState({source: 'geojson', id: 12345});
-                expect(fState.hover).toBe(true);
-                done();
-            });
-        });
-        test('works with string ids', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 'foo'}, {'hover': true});
-                const fState = map.getFeatureState({source: 'geojson', id: 'foo'});
-                expect(fState.hover).toBe(true);
-                done();
-            });
-        });
-        test('parses feature id as an int', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: '12345'}, {'hover': true});
-                const fState = map.getFeatureState({source: 'geojson', id: 12345});
-                expect(fState.hover).toBe(true);
-                done();
-            });
-        });
-        test('throw before loaded', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            expect(() => {
-                map.setFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-            }).toThrow(Error);
-
-            done();
-        });
-        test('fires an error if source not found', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/source/);
-                    done();
-                });
-                map.setFeatureState({source: 'vector', id: 12345}, {'hover': true});
-            });
-        });
-        test('fires an error if sourceLayer not provided for a vector source', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'vector': {
-                            'type': 'vector',
-                            'tiles': ['http://example.com/{z}/{x}/{y}.png']
-                        }
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/sourceLayer/);
-                    done();
-                });
-                (map as any).setFeatureState({source: 'vector', sourceLayer: 0, id: 12345}, {'hover': true});
-            });
-        });
-        test('fires an error if id not provided', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'vector': {
-                            'type': 'vector',
-                            'tiles': ['http://example.com/{z}/{x}/{y}.png']
-                        }
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/id/);
-                    done();
-                });
-                (map as any).setFeatureState({source: 'vector', sourceLayer: '1'}, {'hover': true});
-            });
-        });
-    });
-
-    describe('#removeFeatureState', () => {
-
-        test('accepts "0" id', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 0}, {'hover': true, 'click': true});
-                map.removeFeatureState({source: 'geojson', id: 0}, 'hover');
-                const fState = map.getFeatureState({source: 'geojson', id: 0});
-                expect(fState.hover).toBeUndefined();
-                expect(fState.click).toBe(true);
-                done();
-            });
-        });
-        test('accepts string id', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 'foo'}, {'hover': true, 'click': true});
-                map.removeFeatureState({source: 'geojson', id: 'foo'}, 'hover');
-                const fState = map.getFeatureState({source: 'geojson', id: 'foo'});
-                expect(fState.hover).toBeUndefined();
-                expect(fState.click).toBe(true);
-                done();
-            });
-        });
-        test('remove specific state property', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-                map.removeFeatureState({source: 'geojson', id: 12345}, 'hover');
-                const fState = map.getFeatureState({source: 'geojson', id: 12345});
-                expect(fState.hover).toBeUndefined();
-                done();
-            });
-        });
-        test('remove all state properties of one feature', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 1}, {'hover': true, 'foo': true});
-                map.removeFeatureState({source: 'geojson', id: 1});
-
-                const fState = map.getFeatureState({source: 'geojson', id: 1});
-                expect(fState.hover).toBeUndefined();
-                expect(fState.foo).toBeUndefined();
-
-                done();
-            });
-        });
-        test('remove properties for zero-based feature IDs.', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 0}, {'hover': true, 'foo': true});
-                map.removeFeatureState({source: 'geojson', id: 0});
-
-                const fState = map.getFeatureState({source: 'geojson', id: 0});
-                expect(fState.hover).toBeUndefined();
-                expect(fState.foo).toBeUndefined();
-
-                done();
-            });
-        });
-        test('other properties persist when removing specific property', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 1}, {'hover': true, 'foo': true});
-                map.removeFeatureState({source: 'geojson', id: 1}, 'hover');
-
-                const fState = map.getFeatureState({source: 'geojson', id: 1});
-                expect(fState.foo).toBe(true);
-
-                done();
-            });
-        });
-        test('remove all state properties of all features in source', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 1}, {'hover': true, 'foo': true});
-                map.setFeatureState({source: 'geojson', id: 2}, {'hover': true, 'foo': true});
-
-                map.removeFeatureState({source: 'geojson'});
-
-                const fState1 = map.getFeatureState({source: 'geojson', id: 1});
-                expect(fState1.hover).toBeUndefined();
-                expect(fState1.foo).toBeUndefined();
-
-                const fState2 = map.getFeatureState({source: 'geojson', id: 2});
-                expect(fState2.hover).toBeUndefined();
-                expect(fState2.foo).toBeUndefined();
-
-                done();
-            });
-        });
-        test('specific state deletion should not interfere with broader state deletion', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 1}, {'hover': true, 'foo': true});
-                map.setFeatureState({source: 'geojson', id: 2}, {'hover': true, 'foo': true});
-
-                map.removeFeatureState({source: 'geojson', id: 1});
-                map.removeFeatureState({source: 'geojson', id: 1}, 'foo');
-
-                const fState1 = map.getFeatureState({source: 'geojson', id: 1});
-                expect(fState1.hover).toBeUndefined();
-
-                map.setFeatureState({source: 'geojson', id: 1}, {'hover': true, 'foo': true});
-                map.removeFeatureState({source: 'geojson'});
-                map.removeFeatureState({source: 'geojson', id: 1}, 'foo');
-
-                const fState2 = map.getFeatureState({source: 'geojson', id: 2});
-                expect(fState2.hover).toBeUndefined();
-
-                map.setFeatureState({source: 'geojson', id: 2}, {'hover': true, 'foo': true});
-                map.removeFeatureState({source: 'geojson'});
-                map.removeFeatureState({source: 'geojson', id: 2}, 'foo');
-
-                const fState3 = map.getFeatureState({source: 'geojson', id: 2});
-                expect(fState3.hover).toBeUndefined();
-
-                done();
-            });
-        });
-        test('add/remove and remove/add state', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.setFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-
-                map.removeFeatureState({source: 'geojson', id: 12345});
-                map.setFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-
-                const fState1 = map.getFeatureState({source: 'geojson', id: 12345});
-                expect(fState1.hover).toBe(true);
-
-                map.removeFeatureState({source: 'geojson', id: 12345});
-
-                const fState2 = map.getFeatureState({source: 'geojson', id: 12345});
-                expect(fState2.hover).toBeUndefined();
-
-                done();
-            });
-        });
-        test('throw before loaded', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            expect(() => {
-                (map as any).removeFeatureState({source: 'geojson', id: 12345}, {'hover': true});
-            }).toThrow(Error);
-
-            done();
-        });
-        test('fires an error if source not found', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': createStyleSource()
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/source/);
-                    done();
-                });
-                (map as any).removeFeatureState({source: 'vector', id: 12345}, {'hover': true});
-            });
-        });
-        test('fires an error if sourceLayer not provided for a vector source', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'vector': {
-                            'type': 'vector',
-                            'tiles': ['http://example.com/{z}/{x}/{y}.png']
-                        }
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/sourceLayer/);
-                    done();
-                });
-                (map as any).removeFeatureState({source: 'vector', sourceLayer: 0, id: 12345}, {'hover': true});
-            });
-        });
-        test('fires an error if state property is provided without a feature id', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'vector': {
-                            'type': 'vector',
-                            'tiles': ['http://example.com/{z}/{x}/{y}.png']
-                        }
-                    },
-                    'layers': []
-                }
-            });
-            map.on('load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/id/);
-                    done();
-                });
-                (map as any).removeFeatureState({source: 'vector', sourceLayer: '1'}, {'hover': true});
-            });
-        });
     });
 
     describe('error event', () => {
